+++ CHANGELOG.md
@@ -28,6 +28,7 @@ BUG FIXES:
   - commands/package: base package won't crash with exception [GH-4017]
   - commands/rsync-auto: Destroyed machines won't raise exceptions. [GH-4031]
   - communicators/ssh: Nicer error if remote unexpectedly disconects. [GH-4038]
+  - communicators/ssh: Clean error when max sessions is hit. [GH-4044]
   - communicators/winrm: Support `mkdir` [GH-4271]
   - guests/centos: Fix issues when NFS client is installed by restarting
       NFS [GH-4088]